Move TelegramPollLogger out of view package, as it is not really a view
